+++ src/components/issue/IssueDetail.jsx
@@ -1,0 +1,5 @@
+const IssueDetail = () => {
+  return <div>이슈 디테일 페이지</div>;
+};
+
+export default IssueDetail;
