+++ src/components/ui/Profile.jsx
@@ -1,0 +1,5 @@
+const Profile = () => {
+  return <div>프로필 사진</div>;
+};
+
+export default Profile;
